[33] refactor: separate struct implementations from endpoints
