+++ src/routes/debate.rs
@@ -1,4 +1,4 @@
-use crate::{omni_error::OmniError, setup::AppState, users::{permissions::Permission, TournamentUser}};
+use crate::{omni_error::OmniError, setup::AppState, tournament_impl::debate_impl::{Debate, DebatePatch}, users::{permissions::Permission, TournamentUser}};
 use axum::{
     extract::{Path, State},
     http::{HeaderMap, StatusCode},
@@ -6,103 +6,11 @@ use axum::{
     routing::get,
     Json, Router,
 };
-use serde::{Deserialize, Serialize};
-use serde_inline_default::serde_inline_default;
-use sqlx::{query, query_as, Pool, Postgres};
+use sqlx::query_as;
 use tower_cookies::Cookies;
 use tracing::error;
-use utoipa::ToSchema;
 use uuid::Uuid;
 
-#[derive(Serialize, Deserialize, ToSchema)]
-#[serde(deny_unknown_fields)]
-pub struct Debate {
-    #[serde(skip_deserializing)]
-    #[serde(default = "Uuid::now_v7")]
-    id: Uuid,
-    motion_id: Option<Uuid>,
-    marshall_user_id: Option<Uuid>,
-    tournament_id: Uuid,
-}
-
-#[serde_inline_default]
-#[derive(Deserialize, ToSchema)]
-pub struct DebatePatch {
-    motion_id: Option<Uuid>,
-    marshall_user_id: Option<Uuid>,
-    tournament_id: Option<Uuid>,
-}
-
-impl Debate {
-    pub async fn post(
-        debate: Debate,
-        connection_pool: &Pool<Postgres>,
-    ) -> Result<Debate, OmniError> {
-        match query_as!(
-            Debate,
-            r#"INSERT INTO debates(id, motion_id, marshall_user_id, tournament_id)
-            VALUES ($1, $2, $3, $4) RETURNING id, motion_id, marshall_user_id, tournament_id"#,
-            debate.id,
-            debate.motion_id,
-            debate.marshall_user_id,
-            debate.tournament_id
-        )
-        .fetch_one(connection_pool)
-        .await
-        {
-            Ok(_) => Ok(debate),
-            Err(e) => Err(e)?,
-        }
-    }
-
-    pub async fn get_by_id(
-        id: Uuid,
-        connection_pool: &Pool<Postgres>,
-    ) -> Result<Debate, OmniError> {
-        match query_as!(Debate, "SELECT * FROM debates WHERE id = $1", id)
-            .fetch_one(connection_pool)
-            .await
-        {
-            Ok(debate) => Ok(debate),
-            Err(e) => Err(e)?,
-        }
-    }
-
-    pub async fn patch(
-        self,
-        patch: DebatePatch,
-        connection_pool: &Pool<Postgres>,
-    ) -> Result<Debate, OmniError> {
-        let debate = Debate {
-            id: self.id,
-            motion_id: patch.motion_id,
-            marshall_user_id: patch.marshall_user_id,
-            tournament_id: patch.tournament_id.unwrap_or(self.tournament_id),
-        };
-        match query!(
-            "UPDATE debates SET motion_id = $1, marshall_user_id = $2 WHERE id = $3",
-            debate.motion_id,
-            debate.marshall_user_id,
-            debate.id,
-        )
-        .execute(connection_pool)
-        .await
-        {
-            Ok(_) => Ok(debate),
-            Err(e) => Err(e)?,
-        }
-    }
-
-    pub async fn delete(self, connection_pool: &Pool<Postgres>) -> Result<(), OmniError> {
-        match query!("DELETE FROM debates WHERE id = $1", self.id)
-            .execute(connection_pool)
-            .await
-        {
-            Ok(_) => Ok(()),
-            Err(e) => Err(e)?,
-        }
-    }
-}
 
 pub fn route() -> Router<AppState> {
     Router::new()
@@ -274,8 +182,7 @@ async fn patch_debate_by_id(
     State(state): State<AppState>,
     headers: HeaderMap,
     cookies: Cookies,
-    Path(tournament_id): Path<Uuid>,
-    Path(id): Path<Uuid>,
+    Path((id, tournament_id)): Path<(Uuid, Uuid)>,
     Json(new_debate): Json<DebatePatch>,
 ) -> Result<Response, OmniError> {
     let pool = &state.connection_pool;
